+++ docs/changelog.md
@@ -20,6 +20,12 @@
 ## workspace
 - rename `src/index.rs` to `src/_.rs`.
 
+## features & flags
+- new feature: `unsafe_mmio`.
+
+## documentation
+- new tag: `_TAG_HW`.
+
 ## examples
 - add minimal no_std environment examples.
 
@@ -39,6 +45,12 @@
 ## sys
 - restrict Linux syscall-backed APIs to compatible Linux/freestanding targets, fixing cross-target builds.
 - avoid detecting host native libraries as available when cross-compiling.
+
+### sys::hw
+- make module public.
+
+#### sys::hw::mcu::avr
+- new types: `AvrPort`, `AvrReg8`.
 
 ### sys::mem
 - make `Ptr` provenance-related methods const: